requirements.txt
beautifulsoup4==4.8.0
docker-compose==1.29.2
elasticsearch==7.14.1
selenium==3.141.0
streamlit==1.22.0
protobuf==3.20.3
python-dotenv
